REF: rename recovery_targets to targets
